@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Recipes.Application.Queries;
@@ -25,7 +25,7 @@ namespace Recipes.Api
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddControllers();
             services.AddDbContext<RecipesDbContext>();
             services.AddScoped<IRecipeQueries, RecipeQueries>();
             services.AddScoped<IRecipeMapper, RecipeMapper>();
@@ -41,7 +41,7 @@ namespace Recipes.Api
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -54,8 +54,16 @@ namespace Recipes.Api
             }
 
             app.UseHttpsRedirection();
-            app.UseSpaStaticFiles();
-            app.UseMvc();
+
+            if (!env.IsDevelopment()) {
+                app.UseSpaStaticFiles();
+            }
+
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints => {
+                endpoints.MapControllers();
+            });
 
             app.UseSpa(spa =>
             {
